requirements.txt
ccxt==1.59.61
pandas==1.3.3
plotly==5.1.0
streamlit==1.9.0
python-decouple==3.5
